requirements.txt
# CORE
Django==4.2.7
psycopg2==2.9.9

# REST
djangorestframework==3.14.0
drf-spectacular==0.26.3

# DEBUG
django-debug-toolbar==4.2.0